back/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=1.8.0
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
pytest-asyncio==0.21.1
python-multipart==0.0.5 
requests==2.32.3
pytest-cov==4.1.0
coverage==7.3.2
slowapi==0.1.8
prometheus-client==0.19.0
redis==5.0.1
pydantic-settings==2.1.0